requirements.txt
telethon==1.37.0
python-dotenv==1.0.1
cryptg==0.4.0
httpx==0.27.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
feedparser==6.0.11
python-multipart==0.0.7
